로그삭제

+++ src/api/common/Web-api.ts
@@ -2,6 +2,7 @@
 // 공용 클래스 api  / 환경설정 임포트 / 모델임포트
 import axios from 'axios';
 import {environment} from '@/utill/environment';
+import {Vue} from 'vue-property-decorator';
 
 export default class WebApi<T>{
     ApiUrl : string = environment.apiUrl
@@ -9,19 +10,32 @@ export default class WebApi<T>{
     headers : any = "";
 
     constructor(){
-        axios.interceptors.response.use(response=>{
-                return response
-            },
-            error=>{
-                if(error.status==401 || error.code == 401){
-                    sessionStorage.clear()
-                    window.location.href='/#/login'
-                }
-                else{
-                    // window.location.href='/#/error'
-                }
-                return Promise.reject(error.response);
-            })
+
+        axios.interceptors.response.use((response) => { // intercept the global error
+            return response
+        }, function (error) {
+            let originalRequest = error.config
+            if (error.response.status === 401 || error.response.code === 401 ) { // if the error is 401 and hasent already been retried
+                // originalRequest._retry = true
+                sessionStorage.clear()
+                alert('세션이 만료 되었습니다 로그인으로 이동합니다')
+                window.location.href = '/#/login'
+            }
+            if (error.response.status === 404 || error.response.code === 401 ) {
+                // originalRequest._retry = true
+                sessionStorage.clear()
+                alert('세션이 만료 되었습니다 로그인으로 이동합니다')
+                window.location.href = '/#/login'
+                return
+            }
+            // if (error.response.status === 500 && !originalRequest._retry) {
+            //     originalRequest._retry = true
+            //     window.location.href = '/#/error'
+            //     return
+            // }
+            // Do something with response error
+            return Promise.reject(error)
+        })
     }
 
     newApi(){
